@@ -1,17 +1,17 @@
 module github.com/MontFerret/worker
 
-go 1.17
+go 1.19
 
 require (
-	github.com/MontFerret/ferret v0.16.6
+	github.com/MontFerret/ferret v0.16.7
 	github.com/go-waitfor/waitfor v1.0.0
 	github.com/go-waitfor/waitfor-http v1.0.0
-	github.com/hashicorp/golang-lru v0.5.4
-	github.com/labstack/echo/v4 v4.6.3
+	github.com/hashicorp/golang-lru/v2 v2.0.1
+	github.com/labstack/echo/v4 v4.9.1
 	github.com/namsral/flag v1.7.4-pre
 	github.com/pkg/errors v0.9.1
-	github.com/rs/zerolog v1.26.1
-	github.com/ziflex/lecho/v3 v3.1.0
+	github.com/rs/zerolog v1.28.0
+	github.com/ziflex/lecho/v3 v3.3.0
 )
 
 require (
@@ -27,20 +27,18 @@ require (
 	github.com/golang/groupcache v0.0.0-20200121045136-8c9f03a8e57e // indirect
 	github.com/gorilla/css v1.0.0 // indirect
 	github.com/gorilla/websocket v1.4.2 // indirect
-	github.com/labstack/gommon v0.3.1 // indirect
+	github.com/labstack/gommon v0.4.0 // indirect
 	github.com/mafredri/cdp v0.32.0 // indirect
-	github.com/mattn/go-colorable v0.1.11 // indirect
+	github.com/mattn/go-colorable v0.1.12 // indirect
 	github.com/mattn/go-isatty v0.0.14 // indirect
 	github.com/sethgrid/pester v1.1.0 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasttemplate v1.2.1 // indirect
 	github.com/wI2L/jettison v0.7.4 // indirect
-	golang.org/x/crypto v0.0.0-20211215165025-cf75a172585e // indirect
+	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
 	golang.org/x/net v0.0.0-20211209124913-491a49abca63 // indirect
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
 	golang.org/x/sys v0.0.0-20220318055525-2edf467146b5 // indirect
 	golang.org/x/text v0.3.7 // indirect
 	golang.org/x/time v0.0.0-20201208040808-7e3f01d25324 // indirect
 )
-
-replace github.com/wI2L/jettison v0.7.3 => github.com/wI2L/jettison v0.7.4
